requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
greenlet==3.2.4
playwright==1.55.0
psycopg2-binary==2.9.10
pyee==13.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
